Make get food by id endpoint pass tests
