+++ app.js
@@ -6,7 +6,7 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-var foodsRouter = require('./routes/api/v1/foods');
+var apiV1Router = require('./routes/api/v1');
 
 var app = express();
 var cors = require('cors');
@@ -26,7 +26,7 @@ app.use(cors())
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/api/v1/foods', foodsRouter);
+app.use('/api/v1', apiV1Router);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
