--- routes/api/v1/foods.js
@@ -1,15 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const sequelize = require('../../../sequelize')
-const Food = require('../../../models/food')
-
-/* GET foods listing. */
-router.get('/', function (req, res) {
-  Food(sequelize).findAll()
-    .then(foods => {
-      res.json(foods)
-    })
-    .catch((e) => console.log(e));
-});
-
-module.exports = router;
